Added All Color Support, ColorPicker Initial Functionality And Fixed A Minize Form Bug
All 20 colors are now supported, maybe I will add some more in the future, or I will allow the user to choose their own color. color picker kind of works, I still need to fix some bugs there. Also now the information of the form is stored when the form is minimized and restored when the form becomes normal again.